refactor(server): rename FutureResultRpcCats to FutureResultRpc
