+++ actor-server/actor-core/src/main/scala/im/actor/api/rpc/FutureResultRpc.scala
@@ -2,4 +2,4 @@ package im.actor.api.rpc
 
 import im.actor.concurrent.FutureResult
 
-object FutureResultRpcCats extends FutureResult[RpcError]
+object FutureResultRpc extends FutureResult[RpcError]
