+++ actor-server/actor-core/src/main/scala/im/actor/api/rpc/raw/MapStyleRawApiService.scala
@@ -47,7 +47,7 @@ trait ProductImplicits {
 }
 
 abstract class MapStyleRawApiService(system: ActorSystem) extends RawApiService(system) with ProductImplicits {
-  import im.actor.api.rpc.FutureResultRpcCats._
+  import im.actor.api.rpc.FutureResultRpc._
   import system.dispatcher
 
   type Request
