+++ actor-server/actor-rpc-api/src/main/scala/im/actor/server/api/rpc/RawApiExtension.scala
@@ -3,7 +3,7 @@ package im.actor.server.api.rpc
 import akka.actor._
 import cats.data.Xor
 import im.actor.api.rpc.collections.ApiRawValue
-import im.actor.api.rpc.FutureResultRpcCats
+import im.actor.api.rpc.FutureResultRpc
 import im.actor.api.rpc.raw.RawApiService
 import im.actor.api.rpc.{ AuthorizedClientData, CommonRpcErrors, RpcError }
 
@@ -13,7 +13,7 @@ import scala.concurrent.Future
 sealed trait RawApiExtension extends Extension
 
 private[rpc] final class RawApiExtensionImpl(system: ExtendedActorSystem) extends RawApiExtension {
-  import FutureResultRpcCats._
+  import FutureResultRpc._
   import system.dispatcher
 
   private val services = TrieMap.empty[String, RawApiService]
